Refactor Retry logic into it's own class for potential future additional retry features like video recording for failed tests etc. Also fix scoping for logcatRecorder as scope should be passed down so that when the outer scope ends, the inner one gets ended as well.
